Fix search results on Android TV
Search results were failing to render due to not having the
appropriate leanback theme applied to the search results
activity. Once reapplied, it was found that search results
were not unique. Fixing how the search result db cleaned up
the existing entries resolved this. The latest leanback support
library was also applied, as was setting the search button
color to the defined accent color. Also, a merge issue was
identified which prevented search history from displaying and
was resolved.

fixes #184