Feature/issue 168 update bookmark (#204)
* Update bookmark

This change allows the player, both local and remote on a cast device, to update the playback bookmark on the recording or video. A user could start a video or recording on the device and finish it on frontend or vice a versa.

If an existing bookmark exists, playback will start from the bookmark. However, there is an issue setting the playback position on google cast devices, so, for now, they always start from position 0.

changes #168

* Update list items to reflect HLS status

changes #168

* Fix broken test

changes #168

* Fix all HLS status in lists showing the same status

changes #168