Feature/issue 231 connectivity check (#235)
* prepare for 2.0.3

* Bugfix/issue 122 (#133)

* fixes #122

added back legacy attributes for older android sdk

* fixes #128

* fixes #122

* Fixes #122

* fixes #135 (#136)

* Update versions for release

* Update screenshots for release

* Update changelog for release

* Update version for release

* Update resources for release

* german translation of 2.1.1 changelog (#183)

* Fix search results on Android TV

Search results were failing to render due to not having the
appropriate leanback theme applied to the search results
activity. Once reapplied, it was found that search results
were not unique. Fixing how the search result db cleaned up
the existing entries resolved this. The latest leanback support
library was also applied, as was setting the search button
color to the defined accent color. Also, a merge issue was
identified which prevented search history from displaying and
was resolved.

fixes #184

* Prepare for release

* Prepare for release

* Prepare for release v2.2.1

* Update versions for release

changes #210

* Update screenshots and changelog for release

changes #210

* Update build options for release

changes #210

* Prepare for release

* Prepare for release

* Refactor duplicate code to a common base class

Changes #231

* Add Connection Troubleshooter

The troubleshootter is accessible from the navigation drawer, from the menu when viewing a media item details, or from a snackbar whenever a connection issue is reported.

Changes #231

* Add timing delays and animations to indicate progress

Changes #68

* Add timing delays and animations to indicate progress

Changes #231

* Add Backend Language Verification

MythTV uses UTF-8. It has previously been reported by users that they were having difficulting displaying the data and the result was their character encoding was not in UTF-8. This will help to diagnose this issue.

Fixes #231

* Add minimum mythtv version check

Changes #231

* Update Airplane Mode text

* Reword Troubleshooting text (again.)